added a location to integrate module to verify resume

+++ job/views.py
@@ -7,6 +7,7 @@ from django.contrib.auth.decorators import user_passes_test
 from job.forms import JobForm, JobModelForm
 from job.models import Job, Field, Applicant
 from accounts.models import Company, Address
+from job.resume_verifier import VerifyResume 
             
 # This function is for the authorization
 def job_poster(user):
@@ -47,18 +48,26 @@ def job_detail(request,slug):
         user.save()
         resume = request.FILES['resume']
 
-        # if User applies again to update their credentials
-        if Applicant.objects.filter(job=job, applicant=user):
-            applicant = Applicant.objects.get(job=job, applicant=user)
-            applicant.delete()
-            Applicant.objects.create(job=job, applicant=user, resume=resume)
-            messages.success(request,'Reapplied with updated credentials')
-        
-        # Appling for the first time
-        else:
-            Applicant.objects.create(job=job, applicant=user, resume=resume)
-            messages.success(request,'Job applied Successfully!')
+        resume_status = VerifyResume.resume_verifier()
+        print(resume_status)
+
+        if(resume_status == True):
+
+            # if User applies again to update their credentials
+            if Applicant.objects.filter(job=job, applicant=user):
+                applicant = Applicant.objects.get(job=job, applicant=user)
+                applicant.delete()
+                Applicant.objects.create(job=job, applicant=user, resume=resume)
+                messages.success(request,'Reapplied with updated credentials')
             
+            # Appling for the first time
+            else:
+                Applicant.objects.create(job=job, applicant=user, resume=resume)
+                messages.success(request,'Job applied Successfully!')
+
+        else:
+            messages.error(request, 'Invalid resume')
+
     related_jobs = Job.objects.filter(job_field= job.job_field).exclude(slug=slug)
     context_dict = {'job':job, 'related_jobs': related_jobs}
     return render(request,'job/job-detail.html',context_dict)
